requirement.txt
asgiref==3.2.10
certifi==2020.6.20
Django==3.1
django-elasticsearch-dsl==7.1.4
django-elasticsearch-dsl-drf==0.20.8
django-nine==0.2.3
djangorestframework==3.11.1
elasticsearch==7.9.1
elasticsearch-dsl==7.2.1
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
